ENV VAR AND ROUTE - Changed lower case to upper case ENV vars - Renamed Home to HomeRoute to match code style
